Allow colon as a valid charaster when it's in a path
With the makefile snippet below, make will report the error
"Makefile:5: *** missing separator.  Stop."

 foo=/tmp/bla
 bar=c:/cygwin/tmp/bla

 $(eval $(foo)=1)
 $(eval $(bar)=1)

 all:
    echo "bar=$(bar) foo=$(foo)"

the fix allows the colon in a variable when it's a part of a
valid path (following a character A-Z or a-z and a slash or
backslash fllowed).

the usage of the fix is that Android would like to use the
file/target name as one part of the variable, which leads the
fail of the full build.

Change-Id: I16240e587d203d6a0892d6139c1d4b43af144c0c